+++ assets/js/pages/ResendVerificationEmailPage.vue
@@ -6,9 +6,9 @@
       <div class="col-md-6">
         <div class="card shadow-sm">
           <div class="card-body p-4">
-            <h2 class="card-title mb-4 text-center">Log in</h2>
+            <h2 class="card-title mb-4 text-center">Resend verification email</h2>
 
-            <form :action="loginPath" method="post" @submit="handleSubmit" ref="formRef" novalidate>
+            <form :action="resendPath" method="post" @submit="handleSubmit" ref="formRef" novalidate>
               <input type="hidden" name="_csrf_token" :value="csrfToken">
 
               <div class="mb-3">
@@ -27,25 +27,9 @@
                 </div>
               </div>
 
-              <div class="mb-3">
-                <label for="password" class="form-label">Hasło</label>
-                <input
-                    type="password"
-                    class="form-control"
-                    :class="{ 'is-invalid': v$.password.$error }"
-                    id="password"
-                    name="password"
-                    v-model="form.password"
-                    autocomplete="current-password"
-                >
-                <div class="invalid-feedback" v-for="err in v$.password.$errors" :key="err.$validator">
-                  {{ err.$message }}
-                </div>
-              </div>
-
               <div class="d-grid">
                 <button type="submit" class="btn btn-primary">
-                  Log in
+                  Send verification email
                 </button>
               </div>
             </form>
@@ -59,7 +43,7 @@
             </div>
 
             <div class="mt-3 text-center">
-              <a href="/verify-email/resend">Resend verification email</a>
+              <a href="/login">Login</a>
             </div>
 
           </div>
@@ -68,7 +52,7 @@
 
       <!-- right: img -->
       <div class="col-md-6 d-none d-md-block text-end">
-        <img :src="require('@images/auth/login.png')" alt="Login" class="img-fluid" />
+        <img :src="require('@images/auth/verify.png')" alt="Login" class="img-fluid" />
       </div>
 
     </div>
@@ -78,13 +62,13 @@
 <script setup>
 import { reactive, computed, ref } from 'vue'
 import useVuelidate from '@vuelidate/core'
-import {required, email, minLength, helpers} from '@vuelidate/validators'
+import {required, email, helpers} from '@vuelidate/validators'
 
 const props = defineProps({
   csrfToken: { type: String, required: true },
   lastEmail: { type: String, default: '' },
-  loginPath: { type: String, required: true },
-  errors: { type: String, default: null },
+  resendPath: { type: String, required: true },
+  error: { type: String, default: null },
 })
 
 const formRef = ref(null)
@@ -98,16 +82,12 @@ const rules = {
   email: {
     required: helpers.withMessage('The email is required', required),
     email
-  },
-  password: {
-    required: helpers.withMessage('The password is required', required),
-    minLength: minLength(8)
-  },
+  }
 }
 
 const v$ = useVuelidate(rules, form)
 
-const errorMessage = computed(() => props.errors)
+const errorMessage = computed(() => props.error)
 
 const handleSubmit = async (e) => {
   e.preventDefault()