+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [16.02.2025]
+- Email verification - email service with verification link and verifying email after registration
 - Login and registration with validation and error messages and country selector with store and api service
 - installing axios and configuration for api calls
 ## [16.02.2025]